# contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: ProcureIQ Sales API
  description: API for managing CRM Sales Orders, Products, Accounts, and Notifications in ProcureIQ.
  version: 1.2.0
paths:
...
  /api/v1/fieldservice/resources:
    post:
      summary: Create service resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceResourceRequest'
      responses:
        '201':
          description: Service resource created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceResponse'
  /api/v1/fieldservice/resources/{id}:
    get:
      summary: Retrieve service resource details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service resource retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceResponse'
    put:
      summary: Update service resource details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceResourceRequest'
      responses:
        '200':
          description: Service resource updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceResponse'
    delete:
      summary: Soft-delete service resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service resource deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/v1/fieldservice/skills/{id}:
    get:
      summary: Retrieve skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSkillResponse'
    put:
      summary: Update skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSkillResponse'
    delete:
      summary: Delete skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/v1/fieldservice/service-resource-skills:
    post:
      summary: Create service resource skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceResourceSkillRequest'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceSkillResponse'
  /api/v1/fieldservice/service-resource-skills/{id}:
    get:
      summary: Retrieve service resource skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceSkillResponse'
    put:
      summary: Update service resource skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceResourceSkillRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceSkillResponse'
    delete:
      summary: Delete service resource skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/v1/fieldservice/service-resource-capacities:
    post:
      summary: Create service resource capacity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceResourceCapacityRequest'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceCapacityResponse'
  /api/v1/fieldservice/service-resource-capacities/{id}:
    get:
      summary: Retrieve service resource capacity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceCapacityResponse'
    put:
      summary: Update service resource capacity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceResourceCapacityRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceResourceCapacityResponse'
    delete:
      summary: Delete service resource capacity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
...
    ApiResponseString:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        data:
          type: string
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    ServiceResourceRequest:
      type: object
      required:
        - name
        - resourceType
      properties:
        name:
          type: string
        resourceType:
          type: string
          enum: [technician, crew]
        isActive:
          type: boolean
    ServiceResourceResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        resourceType:
          type: string
        isActive:
          type: boolean
    ApiResponseServiceResourceResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/ServiceResourceResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    SkillRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    SkillResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    ApiResponseSkillResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/SkillResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    ServiceResourceSkillRequest:
      type: object
      required:
        - serviceResourceId
        - skillId
        - skillLevel
        - validFrom
      properties:
        serviceResourceId:
          type: integer
        skillId:
          type: integer
        skillLevel:
          type: integer
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
    ServiceResourceSkillResponse:
      type: object
      properties:
        id:
          type: integer
        serviceResourceId:
          type: integer
        skillId:
          type: integer
        skillLevel:
          type: integer
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
    ApiResponseServiceResourceSkillResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/ServiceResourceSkillResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    ServiceResourceCapacityRequest:
      type: object
      required:
        - serviceResourceId
        - capacityType
        - capacityValue
        - startDate
      properties:
        serviceResourceId:
          type: integer
        capacityType:
          type: string
        capacityValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    ServiceResourceCapacityResponse:
      type: object
      properties:
        id:
          type: integer
        serviceResourceId:
          type: integer
        capacityType:
          type: string
        capacityValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    ApiResponseServiceResourceCapacityResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/ServiceResourceCapacityResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...